LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/input-preprocessing-expected.txt


PASS "foo\0" becomes "foo�"
PASS "f\0oo" becomes "f�oo"
PASS "\0foo" becomes "�foo"
PASS "\0" becomes "�"
PASS "\0\0\0" becomes "���"
FAIL "fooU+d800" becomes "foo�" assert_equals: expected "foo\ufffd" but got "foo�"
FAIL "fU+d800oo" becomes "f�oo" assert_equals: expected "f\ufffdoo" but got "f�oo"
FAIL "U+d800foo" becomes "�foo" assert_equals: expected "\ufffdfoo" but got "�foo"
FAIL "U+d800" becomes "�" assert_equals: expected "\ufffd" but got "�"
FAIL "U+d800U+d800U+d800" becomes "���" assert_equals: expected "\ufffd\ufffd\ufffd" but got "���"
